+++ src/main/java/lotto/model/BuyingType.java
@@ -1,35 +1,33 @@
-package lotto;
+package lotto.model;
 
-import lotto.model.BuyingPocket;
-import lotto.model.Lottos;
 import lotto.type.GenerateType;
 
 import static lotto.Constants.NUMBER_OF_ZERO;
 
-public class LottoBuyingType {
+public class BuyingType {
 
     private Integer randomCount;
     private Integer passiveCount;
     private GenerateType generateType;
 
-    LottoBuyingType(final int randomCount, final int passiveCount, final GenerateType generateType) {
+    BuyingType(final int randomCount, final int passiveCount, final GenerateType generateType) {
         this.randomCount = randomCount;
         this.passiveCount = passiveCount;
         this.generateType = generateType;
     }
 
-    public static LottoBuyingType createBuyingTypeByCounts(final int totalCount, final int passiveCount){
+    public static BuyingType createBuyingTypeByCounts(final int totalCount, final int passiveCount){
         final int randomCount = totalCount - passiveCount;
 
         if(randomCount == NUMBER_OF_ZERO) {
-            return new LottoBuyingType(NUMBER_OF_ZERO, passiveCount, GenerateType.PASSIVE);
+            return new BuyingType(NUMBER_OF_ZERO, passiveCount, GenerateType.PASSIVE);
         }
 
         if(passiveCount == NUMBER_OF_ZERO){
-            return new LottoBuyingType(randomCount, NUMBER_OF_ZERO, GenerateType.RANDOM);
+            return new BuyingType(randomCount, NUMBER_OF_ZERO, GenerateType.RANDOM);
         }
 
-        return new LottoBuyingType(randomCount, passiveCount, GenerateType.BOTH);
+        return new BuyingType(randomCount, passiveCount, GenerateType.BOTH);
     }
 
     public Integer getRandomCount(){
