+++ src/main/java/lotto/service/LottoStore.java
@@ -1,4 +1,7 @@
-package lotto.model;
+package lotto.service;
+
+import lotto.model.BuyingPocket;
+import lotto.model.Lottos;
 
 public class LottoStore {
     public static Lottos getMyLottosByMoney(final BuyingPocket pocket){
